+++ packages/maps/src/base/markers.factory.ts
@@ -1,9 +1,9 @@
-import { MarkerDataConfig, StyleObj } from "../interfaces";
+import { MarkerDataConfig, StyleFunct } from "../interfaces";
 
 export class MarkersBase
   implements MarkerDataConfig<any>, MarkersFactoryConfig {
-  // styleObj: StyleObj;
-  constructor(public renderConfig, public styleObj: StyleObj, public name?) {}
+  // styleFunct: StyleFunct;
+  constructor(public renderConfig, public styleFunct: StyleFunct, public name?) {}
 
   getPosition(data) {
     return this.renderConfig.getPosition(data);
@@ -16,5 +16,5 @@ export class MarkersBase
 
 export interface MarkersFactoryConfig {
   renderConfig: MarkerDataConfig<any>;
-  styleObj: StyleObj;
+  styleFunct: StyleFunct;
 }