+++ packages/maps/src/styles/segment-polyline-styles.ts
@@ -1,17 +1,25 @@
 import { Color } from "ht-utility";
 
 export const segmentPolylineStyles = {
-  google: {
-    default: {
-      strokeColor: Color.blue,
-      strokeOpacity: 1,
-      strokeWeight: 5
+  get(type) {
+    switch (type) {
+      case "google": {
+        return {
+          default: {
+            strokeColor: Color.blue,
+            strokeOpacity: 1,
+            strokeWeight: 5
+          }
+        }
+      };
+      case "leaflet": {
+        return {
+          default: {
+            weight: 5,
+            color: Color.blue
+          }
+        }
+      }
     }
   },
-  leaflet: {
-    default: {
-      weight: 5,
-      color: Color.blue
-    }
-  }
 };